Moved embed video import to a dedicated service. (#8827)
