+++ src/Orchard.Web/Modules/Orchard.MediaLibrary/Services/IOEmbedService.cs
@@ -1,0 +1,7 @@
+﻿using System.Xml.Linq;
+
+namespace Orchard.MediaLibrary.Services {
+    public interface IOEmbedService : IDependency {
+        XDocument DownloadMediaData(string url);
+    }
+}
